reservation controller init
