+++ app/views/garages/show.html.erb
@@ -1,6 +1,6 @@
 <div class="valign-wrapper whole-page">
   <div class="row valign whole-width">
-    <%= form_tag pages_next_url(@booking), :length => params[:length] do %>
+    <%= form_tag reservations_garage_reservation_url(@booking), :length => params[:length] do %>
 
         <div class="input-field">
           <%= label :length, "How long will you stay?" %>
