+++ app/views/pages/home.html.erb
@@ -12,7 +12,7 @@
 
   <a id="reserve-home-btn" class="waves-effect waves-light btn-large">Reserve</a>
   <% if user_signed_in? %>
-    <%= render "pages/reservation" %>
+    <%= render "reservations/home_reservation" %>
   <% end %>
 </div>
