--- app/views/pages/online_reservation.html.erb
@@ -1,19 +1,0 @@
-<center>
-    <h2>Your booking ID: <%= @hash_booking_id %></h2>
-
-    Garage: <%= @booking.garage_spot.garage.name %><br>
-    Spot: <%= @booking.garage_spot.spot.name %><br><br>
-
-    Start Time: <%= @booking.start_time %><br>
-    End Time: <%= @booking.end_time %><br>
-    Length: <%= @booking.length %> hour(s)<br><br>
-
-    Charges: $<%= (@charges.amount)/100 %><br>
-    Paid: <%= @charges.paid %><br><br>
-
-    <div class="qr"><%= raw @qr.as_html %></div>
-
-</center>
-
-<br>
-<%= link_to "print", "#", onclick: "printPage()", class: "button btn btn-info btn-xs" %>
