+++ app/views/reservations/garage_reservation.html.erb
@@ -1,0 +1,1 @@
+<%= render "reservations/get_reservation" %>
